fixed alot did alot dont remember what, yo

+++ client/src/components/LayoutTile.tsx
@@ -12,18 +12,22 @@ export function LayoutTile({layout, onDelete}: LayoutProps){
     
     const layoutTileClassName = 'relative flex flex-col gap-3 justify-between justify-center border p-2 rounded hover:cursor-pointer hover:border-blue-400 hover:shadow-lg hover:text-blue-500 '
 
-    const deleteLayout = async (LID:number|undefined) => {
-        console.log('delete button pressed', LID);
-        await axios.delete('api/users/layouts/deleteLayout/'+LID).catch((err) => {console.log(err);})
-        onDelete()
+    const deleteLayout = async ( event: React.MouseEvent) => {
+        const LID = layout.id
+        event.stopPropagation()
+        if(LID !== undefined){
+            console.log('delete button pressed', LID);
+            await axios.delete('api/users/layouts/deleteLayout/'+LID).catch((err) => {console.log(err);})
+            onDelete()
+        }
     }
     
     return(
         <div className={layoutTileClassName} onClick={() => navigate(`/notes/${layout.id}`)}>
-            <button onClick={() => deleteLayout(layout.id)} className="absolute border bg-red-300 rounded-full right-0 top-0 m-1 p-1 text-black font-bold">Del</button>
-            <h2 className="font-bold">{layout.layout?.title}</h2>
-            <div className="mx-7 px-10 py-1 min-h-[120px] h-[10vw] overflow-hidden text-left border-2 rounded-lg text-black">
-                <ul>
+            <button onClick={deleteLayout} className="absolute border bg-red-300 rounded-full right-0 top-0 m-1 p-1 text-black font-bold hover:text-blue-500">Del</button>
+            <h2 className="font-bold">{layout.title}</h2>
+            <div className="mx-7 px-10 py-1 min-h-[120px] h-[10vw] overflow-hidden text-left border-2 rounded-lg truncate text-black hover:bg-white">
+                <ul className="list-disc">
                     <li>раз</li>
                     <li>два</li>
                     <li>три</li>
@@ -32,6 +36,8 @@ export function LayoutTile({layout, onDelete}: LayoutProps){
                     <li>six</li>
                     <li>seven</li>
                     <li>eight</li>
+                    <li>nine</li>
+                    <li>ten</li>
             </ul></div>
             <button className="border-2 border-blue-200 mx-auto rounded-full p-2 font-bold text-blue-100 bg-blue-500">Edit</button>
         </div>
